in terms of privacy and confidential information.
we have to delete this file google_services.josn because of our api key confidential access.
regarding this i got a mail right now. 
we all have this file in our project and if anyone put changes on this file we will share it manually or other communication channel like skype etc.